refactor: register MixinClientLevel in a new mixin JSON file for NeoForge only
