+++ src/main/resources/epicfight-platform.neoforge.mixins.json
@@ -1,0 +1,12 @@
+{
+  "required": false,
+  "minVersion": "0.8",
+  "compatibilityLevel": "JAVA_21",
+  "package": "yesman.epicfight.platform.neoforge.mixin",
+  "client": [
+    "MixinClientLevel"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
