+++ src/main/templates/META-INF/neoforge.mods.toml
@@ -16,6 +16,9 @@ enumExtensions="META-INF/enum_extensions.json"
 [[mixins]]
 config="${mod_id}.mixins.json"
 
+[[mixins]]
+config="${mod_id}-platform.neoforge.mixins.json"
+
 [[accessTransformers]]
 file="META-INF/accesstransformer.cfg"
